+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This changelog is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ## [unreleased] (Breaking 🐱 | Feat 🚀 | Fixed 🐞)
 
+### Feat 🚀
+
+- Allow to customize release header through `releaseHeaderTemplate` with `{{previousVersion}}` and `{{nextVersion}}` placeholders.
+
 ## [1.0.0] - 2025-10-26
 
 ### Feat 🚀